requirements.txt
google-genai>=1.0.0
yt-dlp>=2024.1.0
python-dotenv>=1.0.0
anthropic>=0.40.0
